requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
ollama>=0.1.0
tqdm>=4.65.0
markdown>=3.4.0
python-dateutil>=2.8.0
numpy>=1.24.0
pillow>=9.0.0
